+++ src/debugger/cordovaDebugSession.ts
@@ -31,6 +31,7 @@ import { CordovaWorkspaceManager } from "../extension/cordovaWorkspaceManager";
 import { CordovaSessionManager } from "../extension/cordovaSessionManager";
 import { SourcemapPathTransformer } from "./cdp-proxy/sourcemapPathTransformer";
 import * as nls from "vscode-nls";
+import { NodeVersionHelper } from "../utils/nodeVersionHelper";
 nls.config({ messageFormat: nls.MessageFormat.bundle, bundleFormat: nls.BundleFormat.standalone })();
 const localize = nls.loadMessageBundle();
 
@@ -440,7 +441,7 @@ export class CordovaDebugSession extends LoggingDebugSession {
             this.cdpProxyLogLevel = args.trace ? LogLevel.Custom : LogLevel.None;
 
             if (args.runtimeVersion) {
-                CordovaProjectHelper.nvmSupport(args);
+                NodeVersionHelper.nvmSupport(args);
             }
         }
     }
